+++ src/app/examples/table-example/table-sticky/table-example.component.ts
@@ -14,7 +14,14 @@ import { sampleData } from '../json-data';
 //ANY, HOUR, COMBOBOX
 export class IbTableStickyExampleComponent implements OnInit {
   @ViewChild('ibTable', {static: true}) ibTable: IbTableComponent;
-  stickyAreas = [IbStickyAreas.HEADER, IbStickyAreas.FOOTER, IbStickyAreas.EDIT, IbStickyAreas.DELETE, IbStickyAreas.SELECT];
+  stickyAreas = [
+    IbStickyAreas.HEADER,
+    IbStickyAreas.FOOTER,
+    IbStickyAreas.EDIT,
+    IbStickyAreas.DELETE,
+    IbStickyAreas.SELECT,
+    IbStickyAreas.SETTINGS
+  ];
   selectableRows = true;
   tableName = 'pippo';
   ibTableActionsPosition = IbTableActionsPosition;
@@ -105,24 +112,7 @@ export class IbTableStickyExampleComponent implements OnInit {
   };
   constructor() { }
 
-  ngOnInit(): void {
-
-    this.ibTable.setFilter('age', [
-      {
-        condition: '>',
-        value: '20'
-      },
-      {
-        condition: '<=',
-        value: '30'
-      }
-    ], 0, true, this.tableName);
-
-    //Full parameters with table name must be sent only when redux is enabled
-    this.ibTable.setFilter('picture', 'place', 0, true, this.tableName);
-
-
-  }
+  ngOnInit(): void {}
 
   consolePrint(item) {
     console.log(item);