+++ src/app/inobeta-ui/ui/table/components/table-header/table-header.component.ts
@@ -1,6 +1,12 @@
 import { Component, OnInit, Input, EventEmitter, Output, HostListener } from '@angular/core';
 import { IbStickyAreas, ibTableSupportedFilters, IbTableTitles, IbTableTitlesTypes } from '../../models/titles.model';
 import { IbTemplateModel } from '../../models/template.model';
+import { Store } from '@ngrx/store';
+import { IbModalMessageService } from '../../../modal/modal-message.service';
+import { IbTableConfSaveComponent } from '../table-conf/table-conf-save.component';
+import { ibTableActionLoadConfig, ibTableActionSaveConfig } from '../../store/actions/table.actions';
+import { IbModalMessage } from '../../../modal/modal-message.model';
+import { IbTableConfLoadComponent } from '../table-conf/table-conf-load.component';
 
 @Component({
   // tslint:disable-next-line: component-selector
@@ -18,7 +24,9 @@ export class IbTableHeaderComponent implements OnInit {
   @Input() columnFilter = {};
   @Input() hasEdit = false;
   @Input() hasDelete = false;
+  @Input() hasConfig = true;
   @Input() stickyAreas = [];
+  @Input() tableName: string;
 
   renderContextMenu = {};
   visibleHeaders = {};
@@ -28,7 +36,10 @@ export class IbTableHeaderComponent implements OnInit {
 
   @Output() handleSetFilter = new EventEmitter<any>();
 
-  constructor() { }
+  constructor(
+    private store: Store<any>,
+    private ibModal: IbModalMessageService
+  ) { }
 
   ngOnInit() {
   }
@@ -61,4 +72,33 @@ export class IbTableHeaderComponent implements OnInit {
   hasCustomHeadersVisible() {
     return Object.keys(this.visibleHeaders).length > 0;
   }
+
+  saveConf() {
+
+    console.log('saveConf');
+    this.ibModal.show({
+      title: 'shared.ibTable.saveConf.title',
+      message: 'shared.ibTable.saveConf.message',
+      tableName: this.tableName
+    } as IbTableConfDialogParams, IbTableConfSaveComponent).subscribe(data => {
+      if (data) {
+        this.store.dispatch(ibTableActionSaveConfig({ options: data, tableName: this.tableName }));
+      }
+    });
+  }
+  loadConf() {
+    this.ibModal.show({
+      title: 'shared.ibTable.loadConf.title',
+      message: 'shared.ibTable.loadConf.message',
+      tableName: this.tableName
+    } as IbTableConfDialogParams, IbTableConfLoadComponent).subscribe(data => {
+      if (data) {
+        this.store.dispatch(ibTableActionLoadConfig({ configName: data.name, tableName: this.tableName }));
+      }
+    });
+  }
+}
+
+export interface IbTableConfDialogParams extends IbModalMessage {
+  tableName: string;
 }