+++ src/app/inobeta-ui/ui/table/components/table-paginator.component.ts
@@ -6,8 +6,8 @@ import {Component, EventEmitter, Input, Output} from '@angular/core';
     <div>
       <mat-paginator
         style="margin-top: 10px;background-color: transparent;"
-        [length]="numOfElements"
-        [pageSize]="elemForPage"
+        [length]="paginationInfo.length"
+        [pageSize]="paginationInfo.pageSize"
         [pageSizeOptions]="[5, 10, 25, 100]"
         [showFirstLastButtons]="true"
         [pageIndex]="paginationInfo.pageIndex"
@@ -20,6 +20,6 @@ export class IbTablePaginatorComponent {
   @Input() numOfElements;
   @Input() paginationInfo;
   @Input() elemForPage = 0;
-
+  
   @Output() pageChangeHandle = new EventEmitter();
 }