+++ src/app/inobeta-ui/ui/table/components/table-total-row/table-total-row-apply-dialog.component.ts
@@ -1,6 +1,6 @@
-import { Component, Inject } from "@angular/core";
-import { MatDialogRef, MAT_DIALOG_DATA } from "@angular/material/dialog";
-import { TotalRowService } from "./total-row.service";
+import { Component, Inject } from '@angular/core';
+import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { TotalRowService } from './total-row.service';
 
 @Component({
   selector: 'ib-table-row-apply-dialog',
@@ -17,7 +17,12 @@ import { TotalRowService } from "./total-row.service";
       </mat-form-field>
     </mat-dialog-content>
     <div mat-dialog-actions align="end">
-      <button mat-button [mat-dialog-close]="{ component: trService.getComponentByFunctionName(function) }">
+      <button mat-button color="warn" [mat-dialog-close]="{
+        func: '_default'
+      }" *ngIf="data.isSet">{{ 'shared.ibTable.func.dialog.remove' | translate }}</button>
+      <button mat-button color="primary" [mat-dialog-close]="{
+        func: function
+      }">
         {{ 'shared.ibTable.apply' | translate }}
       </button>
     </div>
@@ -26,10 +31,11 @@ import { TotalRowService } from "./total-row.service";
 export class IbTableTotalRowApplyDialogComponent {
   function = 'sum';
   functions = Object.entries(TotalRowService.components)
-    .map(([fun]) => ({ value: fun, viewValue: `shared.ibTable.func.${fun}.name`}))
-  
+    .filter(([fun]) => fun !== '_default')
+    .map(([fun]) => ({ value: fun, viewValue: `shared.ibTable.func.${fun}.name`}));
+
   constructor(
-    public trService: TotalRowService,
+    // public trService: TotalRowService,
     public dialogRef: MatDialogRef<IbTableTotalRowApplyDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: any) { }
+    @Inject(MAT_DIALOG_DATA) public data: any) { } 
 }