+++ src/app/inobeta-ui/ui/table/components/table-total-row/table-total-row.component.ts
@@ -2,6 +2,8 @@ import { Component, Input } from '@angular/core';
 import { IbTableItem } from '../../models/table-item.model';
 import { IbTemplateModel } from '../../models/template.model';
 import { IbTableTitles } from '../../models/titles.model';
+import { IbTableTotalRowState } from '../../store/reducers/table.reducer';
+import { TotalRowService } from './total-row.service';
 
 @Component({
   selector: '[ib-table-total-row]',
@@ -16,5 +18,15 @@ export class IbTableTotalRowComponent {
   @Input() hasDelete = false;
   @Input() testData = [];
   @Input() sortedData: IbTableItem[];
-  @Input() filteredData: IbTableItem[]; 
+  @Input() filteredData: IbTableItem[];
+  @Input() totalRowDef: IbTableTotalRowState[];
+  @Input() tableName: string;
+
+  constructor(
+    private totalRowService: TotalRowService
+  ){}
+
+  public getTotalRowCellComponent(key) {
+    return this.totalRowService.getComponentByFunctionName(this.totalRowDef?.find(t => t.columnName === key)?.func);
+  }
 }