+++ src/app/inobeta-ui/ui/table/table.module.ts
@@ -34,6 +34,12 @@ import { StoreModule } from '@ngrx/store';
 import { ibTableFiltersReducer } from './redux/table.reducer';
 import { TotalRowModule } from './components/table-total-row/total-row.module';
 import { IbStickyAreaModule } from './directives/sticky-area/sticky-area.module';
+import { EffectsModule } from '@ngrx/effects';
+import { TableEffects } from './store/effects/table.effects';
+import { ibTableFeatureKey, ibTableFeatureReducer } from './store/reducers/table.reducer';
+import { RouterModule } from '@angular/router';
+import { IbMaterialFormModule } from '../material-forms/material-form.module';
+import { IbTableConfigModule } from './components/table-conf/table-config.module';
 registerLocaleData(localeIt, 'it');
 
 export function ibMatPaginatorTranslate(translateService: TranslateService) {
@@ -41,22 +47,18 @@ export function ibMatPaginatorTranslate(translateService: TranslateService) {
 }
 
 
-const COMPONENTS = [
-  IbTableComponent,
-  IbTableHeaderComponent,
-  IbTableExportComponent,
-  IbTablePaginatorComponent,
-  IbTableHeaderPopupComponent,
-  IbTableExportDialogComponent,
-  IbTableRowComponent,
-  IbTableActionsComponent,
-  IbTableButtonComponent,
-  IbTableHeaderFilterComponent
-];
-
 @NgModule({
   declarations: [
-    ...COMPONENTS,
+    IbTableComponent,
+    IbTableHeaderComponent,
+    IbTableExportComponent,
+    IbTablePaginatorComponent,
+    IbTableHeaderPopupComponent,
+    IbTableExportDialogComponent,
+    IbTableRowComponent,
+    IbTableActionsComponent,
+    IbTableButtonComponent,
+    IbTableHeaderFilterComponent,
     IbTableExportDialogComponent,
   ],
   imports: [
@@ -86,8 +88,24 @@ const COMPONENTS = [
     StoreModule.forFeature('tableFiltersState', ibTableFiltersReducer),
     IbStickyAreaModule,
     TotalRowModule,
+    StoreModule.forFeature(ibTableFeatureKey, ibTableFeatureReducer),
+    EffectsModule.forFeature([TableEffects]),
+    RouterModule,
+    IbTableConfigModule,
   ],
-  exports: [...COMPONENTS],
+  exports: [
+    IbTableComponent,
+    IbTableHeaderComponent,
+    IbTableExportComponent,
+    IbTablePaginatorComponent,
+    IbTableHeaderPopupComponent,
+    IbTableExportDialogComponent,
+    IbTableRowComponent,
+    IbTableActionsComponent,
+    IbTableButtonComponent,
+    IbTableHeaderFilterComponent,
+    IbTableExportDialogComponent,
+  ],
   providers: [
     {
       provide: MatPaginatorIntl, deps: [TranslateService],